Seed 10M MongoDB documents
