+++ package.json
@@ -23,6 +23,7 @@
     "enzyme-adapter-react-16": "^1.12.1",
     "express": "^4.16.4",
     "faker": "^4.1.0",
+    "mongodb": "^3.2.6",
     "mongoose": "^5.4.19",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
